// src/enums/prescription.enum.ts
export enum PrescriptionItemEnum {
    COMPLAINT = "Chief Complaints",
    HISTORY = "History",
    FINDING = "Finding",
    INVESTIGATION = "Investigation",
    MEDICINE = "Medicine",
    ADVICE = "Advice",
    FOLLOW_UP = "Follow-up"
}

export enum MedicineNameTypeEnum {
    GENERIC = "GENERIC",
    MEDICINE = "MEDICINE",
    MEDICINE_WITH_GENERIC = "MEDICINE_WITH_GENERIC"
}

export enum PrescriptionItemListStyleEnum {
    NUMBER = "NUMBER",
    BULLET_POINT = "BULLET_POINT",
    NONE = "NONE"
}

export enum PrescriptionVersionEnum {
    V2 = "v2-desktop",
    V3 = "v3-desktop"
}

export enum SignatureTypeEnum {
    URL = "URL",
    TEXT = "TEXT",
    NOTE = "NOTE"
}